Add support for production Plaid Link flow
This change introduces a new `link-production` command to the Plaid CLI
that facilitates the production Plaid Link flow using a hosted web
experience. It includes handling for a browser-based redirect, launching
the Plaid Link, and capturing the public token via a local server. This
provides a seamless way to integrate Plaid services in production.